Move OptionList into shared components.
