+++ Demo/src/main/java/com/paypal/android/ui/paypalbuttons/PayPalButtonFundingTypeOptionList.kt
@@ -3,7 +3,7 @@ package com.paypal.android.ui.paypalbuttons
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import com.paypal.android.R
-import com.paypal.android.ui.OptionList
+import com.paypal.android.uishared.components.OptionList
 
 @Composable
 fun PayPalButtonFundingTypeOptionList(
